refactor: move highlightText functionality to a utilities
