+++ core/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './pagination'
+export * from './highlight'
